# cmd/main.go
package main

import (
	"fmt"
	"log"
	"net/http"
	"os"

	"github.com/gustionusamba24/kasir-api-go/internal/config"
	"github.com/gustionusamba24/kasir-api-go/internal/controllers"
	"github.com/gustionusamba24/kasir-api-go/internal/repositories/impl"
	serviceImpl "github.com/gustionusamba24/kasir-api-go/internal/services/impl"
	"github.com/joho/godotenv"
	httpSwagger "github.com/swaggo/http-swagger"

	_ "github.com/gustionusamba24/kasir-api-go/docs" // This will be generated by swag init
)

// @title           Kasir API
// @version         1.0
// @description     A simple POS (Point of Sale) API with CRUD operations for categories and products
// @termsOfService  http://swagger.io/terms/

// @contact.name   API Support
// @contact.email  support@kasir-api.com

// @license.name  Apache 2.0
// @license.url   http://www.apache.org/licenses/LICENSE-2.0.html

// @host      localhost:8080
// @BasePath  /

// @schemes http https

func main() {
	// Load .env file - try multiple paths for flexibility
	// This is optional as Railway/production will use environment variables
	if err := godotenv.Load(); err != nil {
		if err := godotenv.Load("../.env"); err != nil {
			if err := godotenv.Load(".env"); err != nil {
				log.Printf("No .env file loaded (this is normal in production)")
			}
		}
	}

	// Connect to the database
	db, err := config.ConnectDatabase()
	if err != nil {
		log.Fatalf("Failed to connect to database: %v", err)
	}
	defer db.Close()

	// Initialize repositories
	categoryRepo := impl.NewCategoryRepository(db)
	productRepo := impl.NewProductRepository(db)
	transactionRepo := impl.NewTransactionRepository(db)

	// Initialize services
	categoryService := serviceImpl.NewCategoryService(categoryRepo)
	productService := serviceImpl.NewProductService(productRepo, categoryRepo)
	transactionService := serviceImpl.NewTransactionService(transactionRepo, productRepo)
	reportService := serviceImpl.NewReportService(transactionRepo)

	// Initialize controllers
	categoryController := controllers.NewCategoryController(categoryService)
	productController := controllers.NewProductController(productService)
	transactionController := controllers.NewTransactionController(transactionService)
	reportController := controllers.NewReportController(reportService)

	// Setup routes
	mux := http.NewServeMux()

	// Category routes
	mux.HandleFunc("/categories", func(w http.ResponseWriter, r *http.Request) {
		switch r.Method {
		case http.MethodGet:
			categoryController.GetAll(w, r)
		case http.MethodPost:
			categoryController.Create(w, r)
		default:
			http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
		}
	})

	mux.HandleFunc("/categories/", func(w http.ResponseWriter, r *http.Request) {
		switch r.Method {
		case http.MethodGet:
			categoryController.GetByID(w, r)
		case http.MethodPut:
			categoryController.Update(w, r)
		case http.MethodDelete:
			categoryController.Delete(w, r)
		default:
			http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
		}
	})

	// Product routes
	mux.HandleFunc("/products", func(w http.ResponseWriter, r *http.Request) {
		switch r.Method {
		case http.MethodGet:
			productController.GetAll(w, r)
		case http.MethodPost:
			productController.Create(w, r)
		default:
			http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
		}
	})

	mux.HandleFunc("/products/", func(w http.ResponseWriter, r *http.Request) {
		switch r.Method {
		case http.MethodGet:
			productController.GetByID(w, r)
		case http.MethodPut:
			productController.Update(w, r)
		case http.MethodDelete:
			productController.Delete(w, r)
		default:
			http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
		}
	})

	// Transaction routes
	mux.HandleFunc("/transactions/checkout", func(w http.ResponseWriter, r *http.Request) {
		if r.Method == http.MethodPost {
			transactionController.Checkout(w, r)
		} else {
			http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
		}
	})

	mux.HandleFunc("/transactions", func(w http.ResponseWriter, r *http.Request) {
		if r.Method == http.MethodGet {
			transactionController.GetAll(w, r)
		} else {
			http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
		}
	})

	mux.HandleFunc("/transactions/", func(w http.ResponseWriter, r *http.Request) {
		if r.Method == http.MethodGet {
			transactionController.GetByID(w, r)
		} else {
			http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
		}
	})

	// Report routes
	mux.HandleFunc("/report/today", func(w http.ResponseWriter, r *http.Request) {
		if r.Method == http.MethodGet {
			reportController.GetTodayReport(w, r)
		} else {
			http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
		}
	})

	mux.HandleFunc("/report", func(w http.ResponseWriter, r *http.Request) {
		if r.Method == http.MethodGet {
			reportController.GetDateRangeReport(w, r)
		} else {
			http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
		}
	})

	// Swagger documentation route
	mux.HandleFunc("/swagger/", httpSwagger.WrapHandler)

	// Root route - Welcome page
	mux.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
		if r.URL.Path != "/" {
			http.NotFound(w, r)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		w.WriteHeader(http.StatusOK)

		port := os.Getenv("PORT")
		if port == "" {
			port = "8080"
		}

		response := fmt.Sprintf(`{
  "message": "Welcome to Kasir API - Point of Sale System",
  "version": "1.0",
  "status": "running",
  "documentation": "http://localhost:%s/swagger/index.html",
  "endpoints": {
    "categories": {
      "getAll": "GET http://localhost:%s/categories",
      "getById": "GET http://localhost:%s/categories/{id}",
      "create": "POST http://localhost:%s/categories",
      "update": "PUT http://localhost:%s/categories/{id}",
      "delete": "DELETE http://localhost:%s/categories/{id}"
    },
    "products": {
      "getAll": "GET http://localhost:%s/products",
      "searchByName": "GET http://localhost:%s/products?name={search_term}",
      "filterByActive": "GET http://localhost:%s/products?active={true|false}",
      "searchWithFilters": "GET http://localhost:%s/products?name={search_term}&active={true|false}",
      "getByCategory": "GET http://localhost:%s/products?category_id={id}",
      "getById": "GET http://localhost:%s/products/{id}",
      "create": "POST http://localhost:%s/products",
      "update": "PUT http://localhost:%s/products/{id}",
      "delete": "DELETE http://localhost:%s/products/{id}"
    },
    "transactions": {
      "checkout": "POST http://localhost:%s/transactions/checkout",
      "getAll": "GET http://localhost:%s/transactions",
      "getById": "GET http://localhost:%s/transactions/{id}"
    },
    "reports": {
      "todayReport": "GET http://localhost:%s/report/today",
      "dateRangeReport": "GET http://localhost:%s/report?start_date={YYYY-MM-DD}&end_date={YYYY-MM-DD}"
    }
  }
}`, port, port, port, port, port, port, port, port, port, port, port, port, port, port, port, port, port, port, port, port)

		fmt.Fprint(w, response)
	})

	// Get port from environment or use default
	port := os.Getenv("PORT")
	if port == "" {
		port = "8080"
	}

	// Start the server
	addr := fmt.Sprintf(":%s", port)
	if err := http.ListenAndServe(addr, mux); err != nil {
		log.Fatalf("Server failed to start: %v", err)
	}
}